feat: finish refactoring of dataset search endpoint
