+++ src/main/java/io/github/genomicdatainfrastructure/discovery/datasets/applications/ports/DatasetsRepository.java
@@ -4,11 +4,14 @@
 
 package io.github.genomicdatainfrastructure.discovery.datasets.applications.ports;
 
-import io.github.genomicdatainfrastructure.discovery.model.DatasetSearchQuery;
 import io.github.genomicdatainfrastructure.discovery.model.SearchedDataset;
 
-import java.util.Set;
+import java.util.List;
 
 public interface DatasetsRepository {
-    Set<SearchedDataset> search(Set<String> datasetIds, String accessToken);
+    List<SearchedDataset> search(List<String> datasetIds,
+                                String sort,
+                                Integer rows,
+                                Integer start,
+                                String accessToken);
 }
