+++ src/main/java/io/github/genomicdatainfrastructure/discovery/datasets/applications/ports/FacetsBuilder.java
@@ -1,10 +1,8 @@
 package io.github.genomicdatainfrastructure.discovery.datasets.applications.ports;
 
 import io.github.genomicdatainfrastructure.discovery.model.DatasetSearchQuery;
-import io.github.genomicdatainfrastructure.discovery.model.DatasetSearchQueryFacet;
-
-import java.util.List;
+import io.github.genomicdatainfrastructure.discovery.model.FacetGroup;
 
 public interface FacetsBuilder {
-    List<DatasetSearchQueryFacet> buildFacets(DatasetSearchQuery query);
+    FacetGroup buildFacets(DatasetSearchQuery query, String accessToken);
 }
