+++ src/main/java/io/github/genomicdatainfrastructure/discovery/datasets/applications/usecases/SearchDatasetsQuery.java
@@ -1,12 +1,13 @@
 package io.github.genomicdatainfrastructure.discovery.datasets.applications.usecases;
 
-import io.github.genomicdatainfrastructure.discovery.datasets.applications.ports.DatasetsRepository;
 import io.github.genomicdatainfrastructure.discovery.datasets.applications.ports.DatasetIdsCollector;
 import io.github.genomicdatainfrastructure.discovery.datasets.applications.ports.FacetsBuilder;
 import io.github.genomicdatainfrastructure.discovery.model.DatasetSearchQuery;
 import io.github.genomicdatainfrastructure.discovery.model.DatasetsSearchResponse;
 import jakarta.inject.Inject;
 
+import java.util.List;
+import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -34,29 +35,32 @@ public class SearchDatasetsQuery {
         var datasetIds = collectors
                 .stream()
                 .map(collector -> collector.collect(this.query, this.accessToken))
+                .filter(Objects::nonNull)
                 .reduce((a, b) -> findIdsIntersection(a, b))
-                .orElse(Set.of());
+                .orElse(List.of());
 
-        var datasets = repository.search(datasetIds, this.accessToken);
+        var datasets = repository.search(datasetIds,
+                                            query.getSort(),
+                                            query.getRows(),
+                                            query.getStart(),
+                                            this.accessToken);
 
-        var facets = facetsBuilders
+        var facetGroups = facetsBuilders
                 .stream()
-                .map(builder -> builder.buildFacets(this.query))
-                .reduce((a, b) -> a.addAll(b) ? a : b)
-                .orElse(Set.of());
+                .map(facetBuilder -> facetBuilder.buildFacets(query, accessToken))
+                .collect(Collectors.toList());
 
         return DatasetsSearchResponse
                 .builder()
-                .count(datasets.size())
-                .results(datasets.stream().toList())
-                .facetGroups(facets)
-                .facetGroupCount(facets.size())
+                .count(datasetIds.size())
+                .results(datasets)
+                .facetGroups(facetGroups)
                 .build();
     }
 
-    private Set<String> findIdsIntersection(Set<String> a, Set<String> b) {
+    private List<String> findIdsIntersection(List<String> a, List<String> b) {
         return a.stream()
                 .filter(b::contains)
-                .collect(Collectors.toSet());
+                .toList();
     }
 }
