+++ src/main/java/io/github/genomicdatainfrastructure/discovery/datasets/infra/quarkus/DatasetsController.java
@@ -1,4 +1,4 @@
 package io.github.genomicdatainfrastructure.discovery.datasets.infra.quarkus;
 
-public class SearchDatasetsController {
+public class DatasetsController {
 }
