--- src/main/java/io/github/genomicdatainfrastructure/discovery/repositories/CkanDatasetsRepository.java
@@ -1,45 +1,0 @@
-// SPDX-FileCopyrightText: 2024 PNED G.I.E.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package io.github.genomicdatainfrastructure.discovery.repositories;
-
-import io.github.genomicdatainfrastructure.discovery.model.DatasetSearchQuery;
-import io.github.genomicdatainfrastructure.discovery.model.DatasetsSearchResponse;
-import io.github.genomicdatainfrastructure.discovery.remote.ckan.api.CkanQueryApi;
-import io.github.genomicdatainfrastructure.discovery.utils.CkanFacetsQueryBuilder;
-import io.github.genomicdatainfrastructure.discovery.utils.PackagesSearchResponseMapper;
-import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.inject.Inject;
-import org.eclipse.microprofile.rest.client.inject.RestClient;
-
-@ApplicationScoped
-public class CkanDatasetsRepository implements IDatasetsRepository {
-
-    private static final String SELECTED_FACETS = "[\"access_rights\",\"theme\",\"tags\",\"spatial_uri\",\"organization\",\"publisher_name\",\"res_format\"]";
-    private final CkanQueryApi ckanQueryApi;
-
-    @Inject
-    public CkanDatasetsRepository(
-            @RestClient CkanQueryApi ckanQueryApi
-    ) {
-        this.ckanQueryApi = ckanQueryApi;
-    }
-
-    @Override
-    public DatasetsSearchResponse search(DatasetSearchQuery query, String accessToken) {
-        var facetsQuery = CkanFacetsQueryBuilder.buildFacetQuery(query);
-
-        var response = ckanQueryApi.packageSearch(
-                query.getQuery(),
-                facetsQuery,
-                query.getSort(),
-                query.getRows(),
-                query.getStart(),
-                SELECTED_FACETS,
-                accessToken
-        );
-
-        return PackagesSearchResponseMapper.from(response).;
-    }
-}
